First commit. Plugin loads, runs ycmd, and gets completions. All calls block.
This uses a more elaborate/lower-level scheme to manage interaction with ycmd.
Different servers are created for different projects, to completers like tern
will be able to inspect the files properly.

The completions are pre-processed and displayed in a normalized way to the
user.  It should hopefully reduce some of the inconsistent descriptions
provided by the third-party completion providers.

All calls and activity happen on the main thread. This will be changed in the
near future... Tasks will be set up to allow execution of certain portions
off-thread, so it won't block the UI.