Converted task pool into a `concurrent.futures.Executor`.
The design of task workers was not very good in terms of having to start and
stop them in client code. Using the design of `concurrent.futures.Executor`
makes it easier for callers to manage the pool, without having to deal with the
individual workers.

Updated unit test to use new pool interface. The setup and teardown code is
significantly reduced while still maintaining all pre-existing functionality.