Added support for parsing diagnostics from completion responses.
These diagnostics are ignored, but they will be used to display errors in the
status bar (for configuration issues), or in the view itself (for things like
linter warnings).

Also added a setting to force semantic completion for all completion requests.
This results in higher quality completions, but may take longer and cause more
timeouts.