Added error diagnostics to detect plugin errors.
This will inspect the loaded settings and display an error dialog if it
detects issues. Most plugin errors will be from this, so it also
includes the reason and a fix.

Cleaned up a lot of error handling to catch and display errors
consistently. Again, this should help with debugging issues.