Add 'Debug Info' and 'Manage Server' commands.
The 'Debug Info' command sends a request to the server for debug info.
This includes information like whether or not an extra config file has
been found/loaded.

The 'Manage Server' command includes two stop commands: Shutdown, which
sends a request and allows the server to exist, and Kill, which
terminates the server process directly. If debug mode is enabled on the
server, the last command may be used to open the corresponding files.